+++ extras/discussion_repl/pyproject.toml
@@ -8,6 +8,7 @@ authors = ["Bob Gardner <bob.hn.gardner@gmail.com>"]
 praw = "^7.6.0"
 python = "^3.10"
 requests = "^2.26.0"
+rich = "^12.4.1"
 
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
