+++ extras/discussion_repl/script/run.py
@@ -6,4 +6,7 @@ import subprocess
 import sys
 
 os.chdir(pathlib.Path(__file__).parents[1])
-subprocess.run(["poetry", "run", "discussion_repl"] + sys.argv[1:], check=True)
+try:
+    subprocess.run(["poetry", "run", "discussion_repl"] + sys.argv[1:], check=True)
+except (KeyboardInterrupt, subprocess.CalledProcessError):
+    sys.exit(1)
